Edit file input function

+++ src/game_life_lib/fileinput.cpp
@@ -14,7 +14,7 @@ void file_input(bool grid[gridsize + 1][gridsize2 + 1], int counter_)
     }
     std::string s1 = "#";
     int* arr = new int(str.length());
-    for (int i = 0; i < str.length(); i++) {
+    for (size_t i = 0; i < str.length(); i++) {
         if (str[i] == s1[0]) {
             arr[i] = 1;
         } else {
